pyproject.toml
[project]
name = "robotframework-asyncio-utils"
version = "0.2.0"
description = "Some keywords to help with async programming in robot framework"
authors = [ {name = "Ygor Pontelo", "email" = "ygorpontelo@gmail.com"} ]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "robotframework>=6.1.1",
]
classifiers = [
    "Framework :: Robot Framework",
]
keywords = ["async", "utils"]

[project.urls]
Homepage = "https://github.com/ygorpontelo/robot-asyncio-utils"

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["AsyncioUtils"]
